Update cc_generate_parallel.py
Initialize mass of new satellites before finding Mlocal. Not doing this can cause Mlocal to be zero for new satellites and, therefore, for M to be used instead of Mlocal. The typical case is when core A falls into core B in one step, and core B falls into core C in the immediate next step: core B's mass would have not been initialized, and therefore core C's halo mass would be used for M.